requirements.txt

pandas>=2.2.2
numpy>=1.26.4
scipy>=1.13.1
matplotlib>=3.8.4
pyyaml>=6.0.2
tabulate>=0.9.0
ib_insync>=0.9.86
